Duplicating Layers into SSLayer
The script now duplicates the contents of each layer into a group inside
the Sprite Sheet Layer. Something important to note is that there is no
way to bring in Effects applied to layers. So a rule for artists will
need to be that they group anything they want to apply effects to and
never apply effects to layers because it won't manifest in the sprite
sheets.

WIP code for now moving those groups into the SS_Artboard. All I need
to do now is make some functions to get the width/height of each group,
and then inteligently organize them into the artboard. Growing the
artboard as necissary.

Lets just plan on hiding all irelivent layers rather than trying to make
the artboard appear in a safe zone. I think this will be a cleaner and
more understandable workflow.