new functions
There are some new helpful functions, most notably are the
resizeAB function which will resize an artboard to a specified width and
height, and the doesIntersect function which checks if to rects
overlap by proving the antithises.

In other words, it's easier to prove that they do not overlap, and if we
can't do that, then it means that they do. While strugling through this
I was beggining to think of this idea, but it wasn't until I saw it
implimented that it confirmed my suspision to proceed with that approach
as a best practice.

I made several modifications however to the link. For one thing, their
approach failed if the boxes were aligned in one axes and only offset in
the other. Mine combines the checks in the return. With this approach,
if I missed a check I can add more test cases easily in the future.