It Works!!!
Finally sorted out the bugs with the movement section. The problem was
that it was checking if it fit AFTER it checked if place free. I needed
to reverse that to so items don't pile up on each other. The second
thing I needed to fix was the check if place free, all the test cases
should be && not || in the final return. Basically, if all tests pass
you get back true, otherwise if a single one of them fails you get back
false.

To help debug things I made a seperate debug(string) function that
relies on a <bool> debug var at the top of the document. This way I can
group my debug alerts in by calling debug instead of alert and then I
can easily turn all debug alerts on/off with that one debugmode var

Also to help with debugging I changed placeFree, doesIntersect, and
depending functions to use the full item objects instead of rects. This
allowed me to completely remove the placedRects array and eliminated a
bunch of silly back and forth from object to rect and back to object.
So now the code is simpler, cleaner and easier to read.

Also, to make that last part work, I went ahead and changed the format
of rectToObject(rect) function to be consistent with the updatePageItem
and updateArtboard functions. All starting with caps X,Y,Left,Top,Right,
Bottom,Width,Height,Size. Before I wanted a distinction but it ended up
just making things more difficult so now they are all consistent.

Just a bit of tidy left to do. It still needs to generate a manafest.
The genSpriteSheet function could probably be split into two seperate
functions one for creating the artboard and layers, and another for
moving all the items into position. Also writing the manafest out to a
file and what to do in the case of a symbol where it's only needed in
the sprite sheet once but should show up in the manafest multiple times.