Require JS
Wrote a simple requireJS(filename) function. This depends on the
Folder.current being set to the location of the calling script. And
assumes that that directory will have a jsRequire folder.

Calling jsRequire() with any file name of a js file inside that
jsRequire folder will cause the global document dom to evaluate that
jsFile before exocuting any other code.

So now I'm able to move the JSON script out to the require folder. And
I should be able to clean up my code and move the helper functions out
to a jsPrototype.js file. Things like forEach etc...